Fixed Player Unit Tests
