+++ MyGameWin.userprefs
@@ -4,29 +4,30 @@
   <MonoDevelop.Ide.DebuggingService.Breakpoints>
     <BreakpointStore />
   </MonoDevelop.Ide.DebuggingService.Breakpoints>
-  <MonoDevelop.Ide.Workbench ActiveDocument="UnitTests\WandererUnitTests.cs">
+  <MonoDevelop.Ide.Workbench ActiveDocument="UnitTests\PlayerUnitTests.cs">
     <Files>
-      <File FileName="LevelEditor.cs" Line="97" Column="21" />
-      <File FileName="Utils.cs" Line="7" Column="2" />
-      <File FileName="WorldAnchor.cs" Line="6" Column="18" />
-      <File FileName="ScreenAnchor.cs" Line="12" Column="35" />
-      <File FileName="Position.cs" Line="61" Column="1" />
-      <File FileName="Renderable.cs" Line="57" Column="43" />
-      <File FileName="Tileset.cs" Line="59" Column="1" />
-      <File FileName="Player.cs" Line="9" Column="48" />
-      <File FileName="Entity.cs" Line="117" Column="16" />
-      <File FileName="TileInteractor.cs" Line="67" Column="1" />
-      <File FileName="UnitTests\RenderableUnitTests.cs" Line="28" Column="9" />
-      <File FileName="UnitTests\RendererUnitTests.cs" Line="6" Column="2" />
-      <File FileName="UnitTests\PatrolUnitTests.cs" Line="1" Column="1" />
-      <File FileName="UnitTests\SpeedBoostTileUnitTests.cs" Line="1" Column="1" />
-      <File FileName="UnitTests\SpeedTileUnitTests.cs" Line="1" Column="1" />
-      <File FileName="UnitTests\WandererUnitTests.cs" Line="9" Column="18" />
+      <File FileName="SpeedTile.cs" Line="10" Column="26" />
+      <File FileName="Tile.cs" Line="7" Column="30" />
+      <File FileName="BitmapRenderable.cs" Line="10" Column="51" />
+      <File FileName="UnitTests\SpeedTileUnitTests.cs" Line="15" Column="10" />
+      <File FileName="Player.cs" Line="12" Column="41" />
+      <File FileName="Enemy.cs" Line="11" Column="45" />
+      <File FileName="UnitTests\PlayerUnitTests.cs" Line="98" Column="4" />
+      <File FileName="UnitTests\WandererUnitTests.cs" Line="35" Column="17" />
+      <File FileName="Entity.cs" Line="17" Column="10" />
+      <File FileName="TileInteractor.cs" Line="46" Column="59" />
+      <File FileName="UnitTests\RenderableUnitTests.cs" Line="14" Column="50" />
+      <File FileName="Renderable.cs" Line="15" Column="16" />
+      <File FileName="Wanderer.cs" Line="51" Column="16" />
+      <File FileName="Tileset.cs" Line="56" Column="52" />
+      <File FileName="Position.cs" Line="58" Column="25" />
     </Files>
     <Pads>
       <Pad Id="MonoDevelop.UnitTesting.TestPad">
         <State name="__root__">
-          <Node name="MyGameWin" expanded="True" selected="True" />
+          <Node name="MyGameWin" expanded="True" selected="True">
+            <Node name="MyGameWin" expanded="True" />
+          </Node>
         </State>
       </Pad>
     </Pads>
